refactor : 지역 선택 패널 로컬 컴포넌트 파일 분리
* 최근 선택 칩·시도 단계·시군구 단계·검색 결과·옵션 행을 형제 파일로 분리
* 옵션 행 공통 props 타입을 OptionRow 모듈로 이동
* 목록 행 공통 스타일 상수를 row-style 모듈로 분리
