+++ apps/web/src/shared/ui/RegionSelect/row-style.ts
@@ -1,0 +1,3 @@
+/** 시·도, 시·군·구 목록 행의 공통 높이·여백. */
+export const ROW =
+  "flex h-[2.4375rem] w-full items-center gap-2 rounded-[0.625rem] px-3.5 text-sm transition";
